fix(dsp): preserve dtype in extract_chunk_segment silence padding (#4134)
The pad branch hardcoded dtype=np.float32 for the silence padding instead of
extracted.dtype, ignoring the buffer's real dtype (the same anti-pattern fixed
in #4125 / chunked_processor). No audible impact today (pipeline is float32
end-to-end) but it violates the dtype-propagation invariant. Now uses
extracted.dtype.

Closes #4134

+++ auralis-web/backend/core/chunk_operations.py
@@ -244,7 +244,9 @@ class ChunkOperations:
                     f"Chunk {chunk_index}: expected 1-D or 2-D audio array, got shape {extracted.shape}"
                 )
             num_channels = extracted.shape[1] if extracted.ndim > 1 else 1
-            padding = np.zeros((padding_needed, num_channels), dtype=np.float32)
+            # Match the buffer's dtype instead of hardcoding float32 (#4134,
+            # same dtype-hygiene fix as #4125) so padding never forces a cast.
+            padding = np.zeros((padding_needed, num_channels), dtype=extracted.dtype)
 
             if extracted.ndim == 1:
                 extracted = extracted[:, np.newaxis]
